Move getEntityTypes function
